+++ workflows_execution/src/epilot/sdkconfiguration.py
@@ -19,9 +19,9 @@ class SDKConfiguration:
     server_idx: int = 0
     language: str = 'python'
     openapi_doc_version: str = '1.0.0'
-    sdk_version: str = '1.46.0'
-    gen_version: str = '2.139.1'
-    user_agent: str = 'speakeasy-sdk/python 1.46.0 2.139.1 1.0.0 epilot-workflows-execution'
+    sdk_version: str = '1.47.0'
+    gen_version: str = '2.142.2'
+    user_agent: str = 'speakeasy-sdk/python 1.47.0 2.142.2 1.0.0 epilot-workflows-execution'
     retry_config: RetryConfig = None
 
     def get_server_details(self) -> tuple[str, dict[str, str]]: